hcn_cms/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
django-admin-thumbnails==0.2.6
django-filter==2.4.0
django-guardian==2.4.0
django-tagging==0.5.0
django==3.2.5
djangorestframework-simplejwt==4.7.2
djangorestframework==3.12.4
gunicorn==20.1.0
pillow==8.2.0
psycopg2-binary==2.9.1
